+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### v0.0.9 (2025-10-15)
 
 - **Added:** Support por Python 3.14.
-- **Added:** Support for checking variables in docstring sections
+- **Added:** Support for checking Variables (`:var`, `:ivar`, `:cvar`, `:vartype`) in docstring sections
 - **Added:** project.urls` section to `pyproject.toml` 
 - **Added:** Support for Module and Class docstring checks.
 - **Fixed:** `DOC006` false positive.